NLP_ML/requirements.txt
scikit-learn
numpy
pdfplumber
python-docx
spacy
